Moved testing page info temporarily to signed in home page
