+++ src/components/Sidenav.js
@@ -14,15 +14,13 @@ import logo from '../assets/logo.svg';
 const loggedInItems =  [
   { name: 'home', label: 'Home', path: '' },
   { name: 'account', label: 'Account', path: 'account' },
-  { name: 'about', label: 'Testing', path: 'about' },
   { name: 'caseload', label: 'Caseload Management', path: 'caseload_management' },
   { name: 'college_list', label: 'College List', path: 'college_list' },
   { name: 'schedule', label: 'Schedule', path: 'schedule' }
 ]
 
 const loggedOutItems = [
-  { name: 'home', label: 'Home', path: '' },
-  { name: 'about', label: 'Testing', path: 'about' }
+  { name: 'home', label: 'Home', path: '' }
 ]
 
 function Sidenav(props) {
